Delete resource-locker-operator deployment for Chart upgrades
The resource-locker-operator chart has various updates which change the
deployment label selectors leaving ArgoCD unable to update the
deployment.

A first try of this used ArgoCD's `Replace=true` sync option, but this
did not work as expected, and ArgoCD was still unable to replace the
Deployment resource.

This commit deploys an ArgoCD Sync job which deletes the Deployment
resource if the new chart version is different to the currently deployed
chart version.