Make component library version-aware
This commit introduces logic in the component library to generate
resource-locker patches which match the deployed resource-locker CRD
version.

Currently, the library only has logic to ensure we only inject the field
`id` for each patch when resource-locker-operator chart v1.x is deployed
by the component.

Note: if the user configures a chart version v1.x and a image v0.x
things are likely to still break.

Fixes #9